src/flask_app/requirements.txt
flask
requests
psycopg2-binary
pandas
reportlab
markdown
beautifulsoup4
python-dotenv